+++ caten_worship/models/__init__.py
@@ -7,6 +7,7 @@ from .users import User
 from .users_profile import UserProfile
 from .mails import mail
 from .login_managers import login_manager
+from flask_migrate import Migrate
 
 
 def init_app(app):
@@ -14,3 +15,4 @@ def init_app(app):
     login_manager.init_app(app)
     db.create_all()
     db.init_app(app)
+    migrate = Migrate(app, db)